server/requirements.txt
python-dotenv
fastapi
uvicorn
semanticscholar
langgraph
langchain-core
langchain
langchain-google-genai
langchain-tavily
langchain-community
wikipedia
langchain_experimental
plotly
matplotlib
numpy
scipy
statsmodels
pandas
scikit-learn
pydantic
langchain-openai
langchain-mistralai
crawl4ai
openpyxl
pyreadstat
pyreadr
aiosqlite
pyreadstat
pyreadr
supabase==2.5.1
google-generativeai==0.8.0
python-multipart==0.0.12
beautifulsoup4==4.12.3
requests==2.31.0
python-pptx==0.6.23
pypdf
python-magic
markdown
markdownify
pycryptodome
trio
docarray